Migrate permissions to dom
